correct denominator for instrument volume fadeout setting
The specs say fadeout/65536 of the original volume is subtracted from
the volume on each tick of the release phase, but they fail to mention
that when fadeout is read from a .xm file it is shifted left by one bit.
So if you don't do that, it's actually subtracting fadeout/32768 each
tick. (It's also a little odd that FT2 will only let you set fadeout as
high as 0xFFF=4095, and Milkytracker has a special "cut" setting that
means fadeout=0x7FFF, which is actually 1 short of instantaneous. If you
load such an instrument in FT2 it bugs out the fadeout slider control.)