fix EDx_env, kind of, by not cutting envs with note
Don't bother calling cutEnvelope from cutNote; we're already silencing
the note with volumeNode anyway, so the envelopes don't matter in this
case.

But there's a more general problem I don't have a good solution for yet,
with calling cancelScheduledValues (as cutEnvelope does) ahead of time
(i.e. when there's a note delay). Left a FIXME note about it in
cutEnvelope.